airflow/requirements.txt
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-apache-kafka==1.4.1
psycopg2-binary==2.9.9
kafka-python==2.0.2
dbt-core==1.8.7
dbt-postgres==1.8.2
sqlalchemy==1.4.53
pydantic==2.9.2